+++ src/Local_Development/Extras.php
@@ -190,20 +190,7 @@ class Extras extends Settings {
 				if ( ! $r['reject_unsafe_urls'] ) {
 					return $r;
 				}
-				$host = parse_url( $url, PHP_URL_HOST );
-				if ( preg_match( '#^(([1-9]?\d|1\d\d|25[0-5]|2[0-4]\d)\.){3}([1-9]?\d|1\d\d|25[0-5]|2[0-4]\d)$#', $host ) ) {
-					$ip = $host;
-				} else {
-					return $r;
-				}
-
-				$parts = array_map( 'intval', explode( '.', $ip ) );
-				if ( 127 === $parts[0] || 10 === $parts[0] || 0 === $parts[0]
-				|| ( 172 === $parts[0] && 16 <= $parts[1] && 31 >= $parts[1] )
-				|| ( 192 === $parts[0] && 168 === $parts[1] )
-				) {
-					$r['reject_unsafe_urls'] = false;
-				}
+				$this->is_localhost() ? $r['reject_unsafe_urls'] = false : true;
 
 				return $r;
 			},